backend/requirements.txt
flask
pypdf2
rank-bm25
openai
python-dotenv
flask-cors
